ReportSlaveId: use correct slaveid/additional data
The Modbus application spec isn't very clear, but the "slave id" field
in ReportSlaveId response is a single byte, then the single byte of
0x00/0xff for status, _then_ the trailing arbitrary data.  Likewise,
the byte count doesn't include either slave id field.

Further, the standard code was passing an extra argument to the execute
method, so it wasn't even getting as far as doing the wrong thing
normally.

This behaviour verified against a Frer and Schneider real world power
meters, and also the behaviour of libmodbus for the length.

Signed-off-by: Karl Palsson <karlp@remake.is>