ReadRegistersResponseBase: Handle parsing of register data lazily.
We add an attribute to store the raw register block data.

When creating a ReadRegistersResponseBase object with existing data, we
put the given register data into 'values' and leave the raw data as a
None object.

When parsing a response, we leave 'registers' as None and put the value
in the raw data block attribute.

If we need the data in one form or the other, we check that we do not
already have this data, then compute its representation as needed.