Do not trigger InvalidMessageRecievedException if incomplete frame received
It is not clear why a while loop was used. Nothing seems to change the buffer within this block.
Function dataReceived of twisted.internet.protocol.Protocol triggers processIncomingPacket everytime data was read and appends this to the
buffer.

Please review the following changes. Especially the self._process calls with the kwarg error=True.

Further questions:
* Do we need to reset the buffer?
* What happens if the client reads only a chunk of the frame and the cable gets unplugged?
  The next time we request new registers we would append the data from the second call to the present buffer which leads to an invalid frame.