Fix resetFrame, modify processIncomingPacket to allow fragments
Function resetFrame() was setting self._header to an empty dict,
causing an exception on the next attempt to access
self._header['len'].

Function processIncomingPacket() no longer resets the frame if a
complete frame is not found. It will leave the data in the buffer
so that subsequent calls can process frames received from multiple
tcp segments.