Loops socket.recv until receive full data or timeout is expired.
socket.recv(size) waits until it gets some data from the host but
not necessarily the entire response that can be fragmented in
many packets.
To avoid the splitted responses to be recognized as invalid messages
and to be discarded, loops socket.recv until full data is received,
or timeout is expired.