Raise exception when modbus unit unexpectedly closes connection
Raise an error when the modbus unit unexpectedly closes the connection
during a receive data operation without returning any data, and log a
warning if it does return some but not all data before closing the
connection. Detecting and handling this here ensures that
ModbusTcpClient._recv doesn't needlessly loop until it times out after
the stream is closed by the peer, and also makes it certain that a
response of b'' from ModbusTcpClient._recv means there was a timeout
and cannot mean that the peer closed the stream, as it could mean
before.

The previous behavior was to not identify the remote end closing the
socket, triggering a timeout, in turn causing the transaction manager
to treat it as the modbus unit returning a response with errors in
it (raising InvalidMessageReceivedException). This will now raise a
ConnectionException, which falls through to the client, bypassing the
retry mechanisms.

Note that https://docs.python.org/3/library/socket.html does not
contain a full description on the socket.recv method; see also pydoc
for socket.socket.recv.