Default to 4K receive buffer size
Change the _sync method so that when no specific size is requested it
fetches up to 4096 bytes instead of 1 byte at a time from the
operating system. It's not clear what the reason was for setting it 1
byte. It was introduced in pymodbus commit:

https://github.com/riptideio/pymodbus/commit/9d2c864bcf80f4491621339a8db3ecd42dbba309

Patch:

https://github.com/riptideio/pymodbus/pull/292/commits/3b490fee80aa7d06be6e88876a721eaaddd3810d

But there's no rationale there for setting it to 1 byte specifically.
While it may have been to handle some obscure operating system with a
poor network stack implementation this should have been documented, as
it is written it comes across as that the implementer didn't
understand how the TCP/stream socket networking API works.

Note that https://docs.python.org/3/library/socket.html does not
contain a full description on the socket.recv method; see also pydoc
for socket.socket.recv.