Improve timeout error message
For TCP/IP when no data is received at all before a timeout the
phrasing "incomplete message received" in error messages in
ModbusTransactionManager._recv is misleading. This wording assertively
states that a message has been received from the modbus unit, and when
no actual data has received this would imply that a message with an
empty payload has been received by the application, which is not
possible with TCP/IP. More likely it's just a timeout, without any
data at all having been received from the modbus unit.