Close connection on no/erroneous message received
Always close connection on no response/error response from modbus
unit. This attempts to resolve issue #538, where the response of a
previously timed out request over synchronized TCP is being returned
as the response to the current request, by leaving the connection open
only if there is no unanswered response.

This commit only touches the code for synchronized clients, and only
seeks to resolve it for TCP based clients, though the situation for
UDP clients should improve as well.