datastore: sparse stores shouldn't need an initial list
Currently, to create a sparse data store, you are required to provide an
initial list/dict.  The first entry is then used as a "default" for the
reset() call, which seems rather at odds with the "sparse" nature.

Instead, use a supplied intial value if provided, and use that to
properly reset the store to _that_ value.  Similarly don't require an
initial list at all, allowing addresses/registers to be filled in via
follow up calls to .setValues()

Example old code: (We want to register 20 addresses at 0x2000}

 # must provide at least one initial value!
 hr = ModbusSparseDataBlock({0x2000: 0x55aa})
 # actually register what we want using flexible lists.
 hr.setValues(0x2000, [0]*20)

Example new code:

 hr = ModbusSparseDataBlock()
 hr.setValues(0x2000, [0]*20)

Example old code using reset:

 hr = ModbusSparseDataBlock({
	0x2000: 0x55,
	0x3000: 0x66
	})
 hr.reset()

At this point, the store contained two registers, 0x2000 and 0x2001,
both containing the value 0x55.  This hardly seems to have been the
intention of .reset() on a Sparse store, but is obvious for the
sequential store.

Fixes: https://github.com/riptideio/pymodbus/issues/566

Signed-off-by: Karl Palsson <karlp@etactica.com>