Do not filter using unit id in the received response (#1076)
processIncomingPacket() can be told to ignore unit ids other than the
specified one or accept all packets by specifying id 0 or 0xff. The
modbus client would extract the unit id from the incoming data packet
and use that for filtering. At best this is a no-op: the incoming unit
id will match itself. At worst it makes us drop responses: if the data
is fragmented each fragment was incorrectly parsed to extract a unit
id.

Remove the incoming unit id check at this stage. This change is
behavior preserving if the response is not fragmented and fixes
riptideio#688 if it is fragmented.