ModbusProtocol: Cleanup handling of local echo
The local echo code lookes for the local echo inside a received datagram.
However, this can be racy, especially with serial ports. It's always
the case that the local echo should start the transaction, but it's
possible for a partial echo to arrive.

Handle partial echos by dropping a part of the sent_buffer and print
an error if the local echo is not actually received when it is expected.