ModbusProtocol: Append to sent_buffer for local echo
When using broadcast packets with the local echo, the echo is not waited for.
If transport_send() gets called again before the echo is received then
`sent_buffer` is overwritten and the echo that might be coming from the
broadcast packet will be forgotten about.

To fix this, always append the expected echo to `sent_buffer`. Thus if there is
an existing echo that has not been received yet it is not forgotten about and
subsequent receives will do the right thing.