TestBasicModbusProtocol: Add tests for broadcast and local echo
Broadcast packets do not wait for a response and thus any local echo
will be stuck in the recieve buffer. Test to ensure a subsequent packet
can be sent after the broadcast and that the local echo is handled
correctly.

Three tests are done: one where the broadcast echo and the subsequent message
echo are returned in two calls to datagram_recieved(), one where the entire
response is in one call, and one where the echo for the message packet
is split over two datagram_received() calls.