Fix broken server examples\
The current examples server codes are not able to communicate back due to
bad framer arguments. The examples through the helper function that parse
the argument only produce a string and is directly passed through the
class arguments causing the framer attribute of the server to be a string
instead of Framer object. This commit fix that problem.