simulator: Fix context single parameter
Fixes an incomplete feature in `Commit:
d32947517087214d28d71cdcbf3c02355526ab68 Feature/simulator addressing`.

The `single` parameter was used incorrectly when ModbusServerContext was
initialized. This caused the simulator to not work correctly when the
configuration "device_id" was used. The context would wrap the dict in
another, having completely incorrect, resulting in using an invalid
object.

Fix it by defining single correctly.

Signed-off-by: Esa Laakso <esa.laakso@fidelix.com>