simulator-api: Implement registers json endpoint
As a part of the ReST API, add endpoint for handling registers. It is
almost a direct copy-paste from the html variant, as to not inroduce
differing behavior between the two. The only exception is that the
Submit actions use a different set of actions, as some of them do not
make sense for json.

The submit dict is endpoint specific here to not cause confusion later
on the rest of the endpoints.

Signed-off-by: Esa Laakso <esa.laakso@fidelix.com>