simulator-api: Add calls_json endpoint
Same as with the registers endpoint, this one is more or less a
copypaste of the calls html endpoint.

Add own submit dict for calls to handle only calls related actions.

The calls_json endpoint handles only the simulation part of the calls.
Monitoring is not supported. The assumption is that the monitoring is
not usable built-in, and without something like WebSocket, the
monitoring would still require an API call.

Signed-off-by: Esa Laakso <esa.laakso@fidelix.com>