simulator-api: Remove generated dict from json
The restapi will not use the generated dict anywhere at the moment, so
remove it as unneeded. To match the type hints, add a type hint for
callable.

Note that it is arguable if the entire table should be moved to the
handle_json function, but for now it is left as is. The table represents
the default paths that the system follows, and could be integral in
understanding the flow of requests.