Update solar.py
According to the name "solar" and the registers and their length listed, this contrib example was meant for Huawei SUN2000 inverters.
I have such an inverter ans the script did not work.
After this adaptation the script works as good as it can and the registers are documented and an log.info shows the value read, if there was no error.