Fix decoding of ReadFileRecordResponse
The `count` increment was incorrect, because `response_length` already
includes the 1 byte for reference type. This moved the pointer too far
ahead, leading to incorrect slicing of record_data.

Also corrected the unit test for this; it seems to be copied from the
Modbus spec example but had a typo (repeated \x05) and did not assert
correctness of the second sub-req in the PDU.