Do not mutate caller's registers in ModbusPDU
Previously, ModbusPDU would assign the `registers` argument to a member
variable and subsequently convert `bytes` to `int`.

The registers argument is typed as `Sequence`, implying that these
values are immutable. However, the assignment and subsequent conversion
could corrupt a caller's data as the conversion affects the original
object reference. It could also potentially raise an error if the
sequence did not allow assignment (tuple, frozenset).

Now, the sequence is iterated through and a new list is created with
converted values (if necessary), leaving the caller's data unchanged.