Fix bad assumption in async_write_registers_mypy
Because the examples are not fully typed, it's not clear to example
writers what types the class attributes actually are.

In this case, `ModbusPDU.registers: list[int]` was being compared for
equality to `regs2: list[bytes]`. This previously worked because
`ModbusPDU` took ownership of the caller's registers and mutated them.
The mutated values were then copied into the response class and compared
against the caller's corrupted values.

Since `registers` should be treated as immutable due to `ModbusPDU`'s
constructor typing the value as `Sequence`, the example has been updated
to compare the list of registers to their converted type (int).